Added TODO file
Its an asciidoc file, so can generate html (suggested that users do so.)

Since I'm just making the file to quickly note some things down, not
putting much effort into it. This means it'll probably stuck around for
a long time, but the project does need some cleanup and the like
already.